requirements-github.txt
beautifulsoup4==4.12.2
requests==2.31.0
transformers==4.35.2
torch==2.1.1
schedule==1.2.0
python-dotenv==1.0.0
trafilatura==1.6.4
lxml==4.9.3
html5lib==1.1
openai==1.3.8
psycopg2-binary==2.9.9
flask==2.3.3
flask-sqlalchemy==3.1.1
boto3==1.34.0